requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
datasets==2.15.0
pandas==2.1.3
python-dotenv==1.0.0 